[Header.html] Fixup the HTML/CSS for the buddy icon so that we can select the top 5 lines of text (and click links up there).
Interestingly, Firefox doesn't render the mockup correctly with this
change, but removing the height directive (or changing it to 67 or
above) makes it not render completely crappily (though the height ends
up wrong). Adium uses webkit which renders this correctly so I'm not
worrying about it.

Also update the mockup to match the new header.